requirements.txt
pytest
selenium
webdriver-manager
pytest-html
allure-pytest